refactor: 폴더 구조 MVC 패턴에 따라 변경
